apps/predictor/requirements.txt
python-dotenv
pandas
numpy
scikit-learn
optuna
mlforecast
coreforecast
matplotlib
statsforecast
ray
lightgbm
xgboost
catboost
fastapi
uvicorn
utilsforecast